perf: frontend resource optimization - code splitting, polling reduction
- Replace 240KB favicon.ico with existing shield.svg (~0.5KB)
- Disable production sourcemaps (saves 6.4MB per build)
- Increase React Query staleTime 5s→30s, add gcTime 5min
- Reduce polling intervals across 15 components (~50% API call reduction)
- Implement route-level code splitting with React.lazy
  - Extract CertificatesPage, WAFPage, LogsPage, SettingsPage to pages/
  - Lazy-load all route components (Dashboard, ProxyHost, etc.)
  - Main JS bundle: 2.1MB → 612KB (71% reduction)
  - Routes load on-demand as separate chunks
- App.tsx: 886 lines → 539 lines
