+++ ui/src/components/ExploitBlockLogs.tsx
@@ -532,7 +532,7 @@ export function ExploitBlockLogs() {
   const { data, isLoading, error, refetch } = useQuery({
     queryKey: ['exploit-block-logs', filter, page, perPage],
     queryFn: () => fetchLogs(page, perPage, filter),
-    refetchInterval: 15000,
+    refetchInterval: 60000,
   });
 
   // Fetch total stats using stats API (accurate COUNT)
@@ -546,7 +546,7 @@ export function ExploitBlockLogs() {
   const { data: totalData } = useQuery({
     queryKey: ['exploit-block-total', baseStatsFilter],
     queryFn: () => fetchLogStats(baseStatsFilter),
-    refetchInterval: 30000,
+    refetchInterval: 60000,
   });
 
   // Get unique rules from rulesData for filter dropdown
