+++ ui/src/main.tsx
@@ -8,7 +8,8 @@ import './i18n'
 const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
-      staleTime: 5000,
+      staleTime: 30000,
+      gcTime: 5 * 60 * 1000,
       retry: 1,
     },
   },
